Introduce folders `shared` and `operations`
