+++ src/jlink.ts
@@ -4,46 +4,10 @@
  * SPDX-License-Identifier: LicenseRef-Nordic-4-Clause
  */
 
-import { fetchIndex } from './jlinkIndex';
-import type { OnUpdate } from './update';
-import { downloadJLink } from './downloadJLink';
-import { installJLink } from './installJLink';
-import { getInstalledJLinkVersion, isValidVersion } from './jLinkVersion';
-
-interface JLinkState {
-    outdated: boolean;
-    installed: boolean;
-    versionToBeInstalled?: string;
-    installedVersion?: string;
-}
-
-export const getVersionToInstall = async ({
-    fallbackVersion,
-    checkOnline = true,
-}: {
-    fallbackVersion?: string;
-    checkOnline?: boolean;
-} = {}): Promise<JLinkState> => {
-    const onlineRecommendedVersion = checkOnline
-        ? (await fetchIndex().catch(() => undefined))?.version
-        : undefined;
-    const versionToBeInstalled = onlineRecommendedVersion ?? fallbackVersion;
-    const installedVersion = await getInstalledJLinkVersion().catch(
-        () => undefined
-    );
-    const installed = !!installedVersion;
-    const outdated =
-        !installed ||
-        !versionToBeInstalled ||
-        !isValidVersion(installedVersion, versionToBeInstalled);
-
-    return {
-        outdated,
-        installedVersion,
-        installed,
-        versionToBeInstalled,
-    };
-};
+import { fetchIndex } from './operations/fetchIndex';
+import type { OnUpdate } from './shared/update';
+import { downloadJLink } from './operations/downloadJLink';
+import { installJLink } from './operations/installJLink';
 
 export const downloadAndSaveJLink = (
     destinationDir: string,
